move websocket logic to a separate file
